requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
google-generativeai>=0.8.6
pydantic>=2.12.5
python-dotenv>=1.2.1
httpx>=0.28.1
jinja2>=3.1.6
sqlalchemy
asyncpg
psycopg2-binary
mcp
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
fastapi-sso
google-api-python-client
google-auth-oauthlib
email-validator
